+++ src/Reporter.h
@@ -44,7 +44,7 @@ public:
 	using WeirdFlowMap = std::map<IPPair, WeirdCountMap>;
 	using WeirdSet = std::unordered_set<std::string>;
 
-	Reporter();
+	Reporter(bool abort_on_scripting_errors);
 	~Reporter();
 
 	// Initialize reporter-sepcific options	that are defined in script-layer.
@@ -262,6 +262,7 @@ private:
 	bool warnings_to_stderr;
 	bool errors_to_stderr;
 	bool after_zeek_init;
+	bool abort_on_scripting_errors = false;
 
 	std::list<std::pair<const Location*, const Location*> > locations;
